buildkernel: Check if the defconfig exist
Before loading the defconfig and building the kernel, first
check if the defconfig provided exist to prevent an explosion

+++ resources/buildkernel.sh
@@ -64,6 +64,12 @@ function buildkernel() {
     fi
   done
 
+  # Check if the defconfig doesn't exist
+  if [ ! -f $kernel_defconfig ]; then
+    echo -e "$RED$BLD   Error:$WHITE Defconfig: '$kernel_defconfig' is missing$RATT"
+    export kernel_build_failed=true; echo " "; return 1
+  fi
+
   # Get config file
   if [ ! -f $kernel_source/.config ]; then
     cp $kernel_source/arch/$kernel_arch/configs/$kernel_defconfig $kernel_source/.config
